requirements.txt
openai
pyyaml
python-dotenv
rich
gtts
pydub
